+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.2.1
 
 * Fixed: #1 should only prompt to open projects and folders not files
+* Fixed: #2 should not raise runtime error if project folder contains special characters
 
 # 0.2.0